Remove default 0 from PO discount fields to prevent wrong amounts
Discount % and Discount ₱ inputs defaulted to "0", so typing a value
without first clearing the field could produce e.g. "50" instead of
"5". They now start empty with a "0" placeholder, matching the
existing Unit Cost field. Cast discount/cost values to float in
normalizeItems() since empty submitted values arrive as "" rather
than null.

+++ app/Http/Controllers/PurchaseOrderController.php
@@ -278,9 +278,9 @@ class PurchaseOrderController extends Controller
             $pair    = $isSet ? $product->pairedProduct : null;
             $qty     = (int) $item['quantity'];
 
-            $discountPercent = $item['discount_percent'] ?? 0;
-            $discountAmount  = $item['discount_amount'] ?? 0;
-            $unitCost        = $item['unit_cost'] ?? 0;
+            $discountPercent = (float) ($item['discount_percent'] ?? 0);
+            $discountAmount  = (float) ($item['discount_amount'] ?? 0);
+            $unitCost        = (float) ($item['unit_cost'] ?? 0);
 
             $netCost = $unitCost * (1 - ($discountPercent / 100));
             if ($qty > 0 && $discountAmount > 0) {
